feat: Enhanced CPU benchmark handling in system forms - Added auto-population of CPU benchmark scores, API endpoint for retrieving benchmarks, changed benchmark field to integer type, updated dashboard tab persistence, updated changelog to version 1.0.81

+++ app/forms.py
@@ -70,7 +70,7 @@ class CPUForm(FlaskForm):
         DataRequired(),
         NumberRange(min=1, max=128, message="Core count must be between 1 and 128")
     ])
-    benchmark = FloatField('Benchmark Score', validators=[
+    benchmark = IntegerField('Benchmark Score', validators=[
         Optional(),
         NumberRange(min=0, message="Benchmark score must be a positive number")
     ])
